fix get user and fetch user on header

+++ ui/src/app/hooks/use-token.tsx
@@ -1,31 +1,46 @@
 'use client'
 
-import { useState } from "react";
+import { useState, useEffect } from "react";
 
 const tokenKey = 'token';
 
 export default function useToken() {
   const [token, setTokenState] = useState<string | null>(null);
 
-  const getToken = () => {
-    const token = localStorage.getItem(tokenKey);
-    if (token) {
-      setTokenState(token);
+  useEffect(() => {
+    const storedToken = localStorage.getItem(tokenKey);
+    if (storedToken) {
+      setTokenState(storedToken);
     }
-    return token;
+  }, []);
+
+  const getToken = () => {
+    if (typeof window !== 'undefined') {
+      const storedToken = localStorage.getItem(tokenKey);
+      if (storedToken) {
+        setTokenState(storedToken);
+      }
+      return storedToken;
+    }
+    return null;
   }
 
-  const setToken = (token: string) => {
-    localStorage.setItem(tokenKey, token);
-    setTokenState(token);
+  const setToken = (newToken: string) => {
+    if (typeof window !== 'undefined') {
+      localStorage.setItem(tokenKey, newToken);
+      setTokenState(newToken);
+    }
   }
 
   const logout = () => {
-    localStorage.clear();
+    if (typeof window !== 'undefined') {
+      localStorage.clear();
+      setTokenState(null);
+    }
   }
 
   return {
-    token: !token ? getToken() : token,
+    token: token || getToken(),
     isAuthenticated: !!token,
     setToken,
     logout
